LPS-58623 We should keep other users membership(users come from ladp server2 or create in portal) of usergroup when synchronize ldap1 server

+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/exportimport/LDAPUserImporterImpl.java
@@ -1128,7 +1128,9 @@ public class LDAPUserImporterImpl implements LDAPUserImporter, UserImporter {
 			userGroupId);
 
 		for (User user : userGroupUsers) {
-			if (!newUserIds.contains(user.getUserId())) {
+			if (!newUserIds.contains(user.getUserId()) &&
+				(ldapServerId == user.getLdapServerId())) {
+
 				UserLocalServiceUtil.deleteUserGroupUser(
 					userGroupId, user.getUserId());
 			}
